tax rework

+++ database/migrations/2024_07_28_082227_create_bazar_tax_rates_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('bazar_tax_rates', static function (Blueprint $table): void {
             $table->id();
             $table->string('name');
-            $table->integer('value')->unsigned();
+            $table->float('value')->unsigned();
             $table->boolean('shipping')->default(false);
             $table->timestamps();
         });
